refactor(neon_framework,neon_files,neon_news,neon_notes): refactor dialog structure
