+++ packages/neon/neon_files/lib/src/widgets/actions.dart
@@ -2,9 +2,9 @@ import 'package:filesize/filesize.dart';
 import 'package:flutter/material.dart';
 import 'package:neon_files/l10n/localizations.dart';
 import 'package:neon_files/src/blocs/files.dart';
-import 'package:neon_files/src/dialogs/choose_folder.dart';
 import 'package:neon_files/src/models/file_details.dart';
 import 'package:neon_files/src/pages/details.dart';
+import 'package:neon_files/src/widgets/dialog.dart';
 import 'package:neon_framework/platform.dart';
 import 'package:neon_framework/utils.dart';
 import 'package:nextcloud/webdav.dart';
