+++ packages/neon/neon_news/lib/src/widgets/feed_floating_action_button.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:neon_news/l10n/localizations.dart';
 import 'package:neon_news/src/blocs/news.dart';
-import 'package:neon_news/src/dialogs/add_feed.dart';
+import 'package:neon_news/src/widgets/dialog.dart';
 
 class NewsFeedFloatingActionButton extends StatelessWidget {
   const NewsFeedFloatingActionButton({
