+++ packages/neon_framework/lib/src/pages/app_implementation_settings.dart
@@ -7,7 +7,7 @@ import 'package:neon_framework/src/settings/widgets/option_settings_tile.dart';
 import 'package:neon_framework/src/settings/widgets/settings_category.dart';
 import 'package:neon_framework/src/settings/widgets/settings_list.dart';
 import 'package:neon_framework/src/theme/dialog.dart';
-import 'package:neon_framework/src/utils/confirmation_dialog.dart';
+import 'package:neon_framework/src/utils/dialog.dart';
 
 @internal
 class AppImplementationSettingsPage extends StatelessWidget {
