+++ packages/neon_framework/lib/utils.dart
@@ -1,10 +1,9 @@
 export 'package:neon_framework/l10n/localizations.dart';
 export 'package:neon_framework/src/utils/app_route.dart';
-export 'package:neon_framework/src/utils/confirmation_dialog.dart';
+export 'package:neon_framework/src/utils/dialog.dart';
 export 'package:neon_framework/src/utils/exceptions.dart';
 export 'package:neon_framework/src/utils/findable.dart';
 export 'package:neon_framework/src/utils/hex_color.dart';
 export 'package:neon_framework/src/utils/provider.dart';
-export 'package:neon_framework/src/utils/rename_dialog.dart';
 export 'package:neon_framework/src/utils/request_manager.dart' hide Cache;
 export 'package:neon_framework/src/utils/validators.dart';
