+++ packages/neon_framework/lib/widgets.dart
@@ -1,5 +1,5 @@
 export 'package:neon_framework/src/widgets/custom_background.dart';
-export 'package:neon_framework/src/widgets/dialog.dart';
+export 'package:neon_framework/src/widgets/dialog.dart' hide NeonAccountSelectionDialog;
 export 'package:neon_framework/src/widgets/error.dart';
 export 'package:neon_framework/src/widgets/image.dart';
 export 'package:neon_framework/src/widgets/linear_progress_indicator.dart';
